infra/modules/apiauth/lambda/requirements.txt
# Generated by create-requirements.sh; DO NOT EDIT
# Date: Thu 31 Aug 18:01:33 BST 2023
attrs==23.1.0
aws-lambda-powertools==2.23.1
aws-lambda-typing==2.18.0
aws-xray-sdk==2.12.0
boto3==1.28.38
botocore==1.31.38
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
environ-config==23.2.0
idna==3.4
jmespath==1.0.1
pycparser==2.21
PyJWT==2.8.0
pynamodb==5.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.16
wrapt==1.15.0
